moving schema manager integration tests into their own class
